Combine Client and Node class into Node.
